Move api mappers to data
