+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/source/functions/mappers/SignIn.kt
@@ -1,0 +1,8 @@
+package com.gdavidpb.tuindice.data.source.functions.mappers
+
+import com.gdavidpb.tuindice.base.domain.model.SignIn
+import com.gdavidpb.tuindice.data.source.functions.responses.SignInResponse
+
+fun SignInResponse.toSignIn() = SignIn(
+	token = token
+)
